Do not convert Array to Hash when publishing
This changes the behaviour a bit, so a major version bump is needed
for this (although to my knowledge we have ever used the Array->Hash
conversion in any of our projects).

Had to use `.kind_of?(Array)` instead of `.respond_to?(:to_a)` since
ruby Hashes responds to `to_a` and we don't want to convert all Hashes
to Arrays.

close #71