Add class for publishing delayed messages
This is the same as a DefaultExchangePublisher, with the difference
that it sets up a separate `.delayed` queue on which it deploys the
messages. RabbitMQ then takes care of the delaying, thanks to a few
arguments given when declaring the queue, and the `expiration` field
set on each message.

See:
* https://www.cloudamqp.com/docs/delayed-messages.html
  * This solution sets "message ttl" on the queue, while I opted to
    set this per message instead, which makes it easier to change
    dynamically.
* https://www.rabbitmq.com/ttl.html#per-message-ttl-in-publishers