Improve message delaying class
Now it's possible to delay messages for both exchanges and a specific
queue (using routing key). For our normal publishers we have two
separate classes, but right now there's only one class for delating
messages.
Reason: Delaying a messages, regardless of what the message's final
destination is (exchange, queue), it will be published to the default
exchange to the delay queue. It's only the message headers that differs.
Therefore it felt like using two separate classes/modules would complicate
things, although it would remove the if-check...