Move DelayPublisher into DefaultExchangePublisher class
Moved the queue creation code into a separate module. Called it
Utilities instead of just Queue, because if we later move the
queue creation code from the subscriber here, we need to be able
to create exchanges too, not just queues.

At the moment the Utilities class doesn't really add much, but at
least the call to "channel.queue" is out of the publisher class..