Add more tests for Utilities.create_queue
There is no other way in bunny to get the queue, than to check the
Bunny::Queue object that's returned from Channel#queue.

The Twingly::AMQP::Utilities.create_queue method is just meant to
be an internal method, so I think it's fine that it returns the
created queue (even though it isn't usable since the associated
channel is closed).

https://github.com/twingly/twingly-amqp/pull/93#discussion_r565288089