Use quorum queues by default
As we will most likely only use quorum queues (currently we at least
only have queues that will benefit by becoming quorum queues), it
makes sense to use this as the default. This way we don't have to
specify the queue type argument in all of our projects.

Wrote a bit about it here as well: https://github.com/twingly/twingly-amqp/pull/100/files#r1350374071

One unexpected thing that took a while to figure out was that the
specs for `Subscriber#cancel!` didn't ack their messages. This
caused RabbitMQ to say that there were still a consumer for the queue,
even though it had been cancelled. I checked the RabbitMQ Management UI,
and it said the same thing: 1 consumer (although the list of consumers was
empty).
The `Subscriber#raw_queue.consumer_count` method we call in those specs,
gets the numbers directly from RabbitMQ, see [Bunny::Queue#consumer_count][1],
so it's not a quirk in Bunny, it's apparently how RabbitMQ behaves when using
quorum queues.

[1]: https://github.com/ruby-amqp/bunny/blob/80a8fc7aa0cd73f8778df87ae05f28c443d10c0d/lib/bunny/queue.rb#L350-L353