Add channel error callback to Subscription
Introduced an on_error callback in Twingly::AMQP::Subscription to handle channel errors, with a default behavior of raising an exception if not set. Updated tests to cover the new callback and changed the Bunny dependency to require version >= 2.23.0.

In 2.23.0 https://github.com/ruby-amqp/bunny/releases/tag/2.23.0 `#on_error` is invoked when channel closes because of a delivery acknowledgement timeout